+++ src/Modules/Plugins/Main.php
@@ -10,6 +10,8 @@
 
 namespace Faster_Updates\Modules\Plugins;
 
+use Faster_Updates\Modules\Plugins\Upgrader as Plugins_Upgrader;
+
 /*
  * Exit if called directly.
  */
@@ -17,11 +19,11 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-require_once __DIR__ . '/class-upgrader.php';
+new Plugins_Upgrader();
 
 /**
  * Plugins Module: Main.
- * 
+ *
  * @since 1.0.0
  */
 class Main {
@@ -93,9 +95,9 @@ class Main {
 		wp_localize_script(
 			'updates',
 			'_wpUpdatesItemCounts',
-			array(
+			[
 				'totals' => wp_get_update_data(),
-			)
+			]
 		);
 
 		require_once ABSPATH . 'wp-admin/admin-footer.php';
@@ -112,7 +114,7 @@ class Main {
 	 * @return void
 	 */
 	public function bulk_update() {
-		if ( ! current_user_can( 'update_plugins' ) ){
+		if ( ! current_user_can( 'update_plugins' ) ) {
 			wp_die( __( 'Sorry, you are not allowed to update plugins for this site.' ) );
 		}
 
@@ -123,7 +125,7 @@ class Main {
 		} elseif ( isset( $_POST['checked'] ) ) {
 			$plugins = (array) $_POST['checked'];
 		} else {
-			$plugins = array();
+			$plugins = [];
 		}
 
 		$plugins = array_map( 'urldecode', $plugins );