+++ CHANGES.md
@@ -1,3 +1,4 @@
 [unreleased]
+* add `composer.json` and use autoloader
 * add generic hooks for overriding update processing for `update-core.php`
 * initial pass